Save modem parameters to profile 0 and 1.
The CNMI setting appears to be lost when the modem is power cycled.
This does not appear to happen when the PFM enters and leaves
low power mode. Save the CNMI settings to profile 0 and profile 1
to make this setting persistent. This was tested on a PFM unit
and found to prevent the setting from becoming lost.